+++ CHANGELOG.md
@@ -9,6 +9,15 @@ Minor releases include updated stdlib stubs from typeshed.
 
 Pyxel doesn't use SemVer anymore, since most minor releases have at least minor backward incompatible changes.
 
+## 2.6.2 / ????-??-??
+
+This release brings a bug fix since version 2.6.1
+
+### Models
+* Fix bug in model [`photon_collection.load_image`](https://esa.gitlab.io/pyxel/doc/stable/references/model_groups/photon_collection_models.html#load-image) where the photon array is now correctly "added to"
+instead of "replaced by a new one".
+  (See TODO: create issue).
+
 ## 2.6.1 / 2024-10-14
 
 This release brings a number of bug fixes since version 2.6